Rewrite tests for @testing-library/react
This will rewrite the existing tests for Editable and move them into a own file.
